+++ llm/settings.py
@@ -1,0 +1,5 @@
+import os
+
+
+LOCAL_MODELS_DIR = os.getenv("LOCAL_MODELS_DIR", "models/")
+SATURNFS_MODELS_DIR = os.getenv("SATURNFS_MODELS_DIR", None)
