+++ llm/training/medcuna_sft.py
@@ -1,7 +1,11 @@
+"""
+Train Vicuna on PubmedQA dataset
+"""
+
 import os
 import datetime
 import pathlib
-from typing import Any, Dict, List, Tuple
+from typing import Tuple
 
 from datasets import load_dataset
 import torch
@@ -9,31 +13,29 @@ from peft import LoraConfig, PeftModel, get_peft_model, prepare_model_for_kbit_t
 from peft.utils import get_peft_model_state_dict
 from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig, DataCollatorForLanguageModeling, PreTrainedTokenizerBase, Trainer, TrainingArguments
 
-from llm.qa.prompts import ZERO_SHOT
+from llm import settings
+from llm.training.data import LazySupervisedFineTuning, process_pubmed_qa
 
 
-NAME = "medcuna-13b"
-BASE_MODEL = "lmsys/vicuna-13b-v1.3"
+NAME = "medcuna-7b"
+BASE_MODEL = "lmsys/vicuna-7b-v1.3"
 MODEL_MAX_LENGTH = 2048
 TRAIN_DATASET = {
     "path": "pubmed_qa",
     "name": "pqa_unlabeled",
-    "split": "train[:6000]",
+    "split": "train[:-100]",
 }
 EVAL_DATASET = {
     "path": "pubmed_qa",
     "name": "pqa_unlabeled",
-    "split": "train[-20:]",
+    "split": "train[-100:]",
 }
-IGNORE_TOKEN_ID = -100
 DEBUGGING = True
+RESUME_FROM_CHECKPOINT = True
 
 WORLD_SIZE = int(os.getenv("WORLD_SIZE", 1))
 LOCAL_RANK = int(os.getenv("LOCAL_RANK", 0))
 GLOBAL_RANK = int(os.getenv("RANK", 0))
-TRAINING_DIR = os.path.dirname(__file__)
-LOCAL_MODELS_DIR = os.getenv("LOCAL_MODELS_DIR", "models/")
-SATURNFS_MODELS_DIR = os.getenv("SATURNFS_MODELS_DIR", None)
 
 
 def has_checkpoint(dir: str) -> bool:
@@ -52,9 +54,9 @@ def load_base_model(lora_config: LoraConfig, gradient_checkpointing: bool = True
     )
     tokenizer.pad_token = tokenizer.unk_token
 
-    # Load model to GPU specified by deepspeed (if applicable), otherwise use auto
+    # Load model to GPU specified by LOCAL_RANK
     device_map = (
-        {"": LOCAL_RANK} if WORLD_SIZE != 1 else "auto"
+        {"": LOCAL_RANK}
     )
     model = AutoModelForCausalLM.from_pretrained(
         BASE_MODEL,
@@ -82,74 +84,15 @@ def load_base_model(lora_config: LoraConfig, gradient_checkpointing: bool = True
     return model, tokenizer
 
 
-def process_data(
-    qa_example: Dict[str, Any],
-    tokenizer: PreTrainedTokenizerBase,
-) -> Dict:
-    roles = ["Question", "Answer"]
-    context: Dict[str, Any] = qa_example["context"]
-    question: str = qa_example["question"]
-    answer: str = qa_example["long_answer"]
-
-    prompt = ZERO_SHOT.render(
-        context["contexts"], roles=roles, question=question, answer=answer
-    )
-
-    # Tokenize conversations
-    input_ids: torch.Tensor = tokenizer(
-        prompt,
-        return_tensors="pt",
-        padding="max_length",
-        max_length=tokenizer.model_max_length,
-        truncation=True,
-    ).input_ids[0]
-    target = input_ids.clone()
-
-    # Mask targets. Only want to train on responses to the questions, not on how
-    # to generate questions/contexts.
-    sep = f"\n{roles[1]}: "
-    split = prompt.split(sep)
-    assert len(split) == 2
-    split[0] += sep
-
-    instruction_len = len(tokenizer(split[0], add_special_tokens=False).input_ids)
-    target[:instruction_len] = IGNORE_TOKEN_ID
-
-    return dict(
-        input_ids=input_ids,
-        labels=target,
-        attention_mask=input_ids.ne(tokenizer.pad_token_id),
-    )
-
-
-class LazySupervisedDataset(torch.utils.data.Dataset):
-    """Dataset for supervised fine-tuning."""
-
-    def __init__(self, raw_data, tokenizer: PreTrainedTokenizerBase, cache_data: bool = False):
-        super(LazySupervisedDataset, self).__init__()
-        self.tokenizer = tokenizer
-        self.raw_data = raw_data
-        self.cached_data_dict = {}
-
-    def __len__(self):
-        return len(self.raw_data)
-
-    def __getitem__(self, i) -> Dict[str, torch.Tensor]:
-        if i in self.cached_data_dict:
-            return self.cached_data_dict[i]
-
-        ret = process_data(self.raw_data[i], self.tokenizer)
-        self.cached_data_dict[i] = ret
-        return ret
-
-
 def tune():
-    output_dir = os.path.join(LOCAL_MODELS_DIR, NAME)
-    dataset = load_dataset(**TRAIN_DATASET)
+    output_dir = os.path.join(settings.LOCAL_MODELS_DIR, NAME)
+    train_dataset = load_dataset(**TRAIN_DATASET)
+    eval_dataset = load_dataset(**EVAL_DATASET)
 
     training_args = TrainingArguments(
         output_dir,
         do_train=True,
+        do_eval=True,
         bf16=True,
         optim="adamw_torch_fused",
         lr_scheduler_type="cosine",
@@ -157,11 +100,14 @@ def tune():
         gradient_accumulation_steps=8,
         num_train_epochs=3,
         save_strategy="epoch",
+        evaluation_strategy="epoch",
+        eval_accumulation_steps=16,
         learning_rate=2e-5,
         warmup_ratio=0.001,
         logging_steps=1,
         gradient_checkpointing=True,
         ddp_find_unused_parameters=False,
+        report_to="none",
     )
     lora_config = LoraConfig(
         r=8,
@@ -173,24 +119,26 @@ def tune():
     )
     model, tokenizer = load_base_model(lora_config, gradient_checkpointing=training_args.gradient_checkpointing)
     data_collator = DataCollatorForLanguageModeling(tokenizer, mlm=False)
-    supervised_dataset = LazySupervisedDataset(dataset, tokenizer)
+    lazy_training = LazySupervisedFineTuning(train_dataset, tokenizer, process_data=process_pubmed_qa)
+    lazy_eval = LazySupervisedFineTuning(eval_dataset, tokenizer, process_data=process_pubmed_qa)
 
     trainer = Trainer(
         model,
         args=training_args,
         data_collator=data_collator,
-        train_dataset=supervised_dataset,
+        train_dataset=lazy_training,
+        eval_dataset=lazy_eval,
         tokenizer=tokenizer,
     )
-    trainer.train(resume_from_checkpoint=has_checkpoint(output_dir))
+    trainer.train(resume_from_checkpoint=RESUME_FROM_CHECKPOINT and has_checkpoint(output_dir))
 
     state_dict = get_peft_model_state_dict(model)
     if GLOBAL_RANK == 0:
         model.save_pretrained(output_dir, state_dict=state_dict)
-        if SATURNFS_MODELS_DIR:
+        if settings.SATURNFS_MODELS_DIR:
             from saturnfs import SaturnFS
             timestamp = datetime.datetime.utcnow().strftime("%Y-%m-%dT%H.%M.%S")
-            remote_dir = os.path.join(SATURNFS_MODELS_DIR, f"{NAME}-{timestamp}")
+            remote_dir = os.path.join(settings.SATURNFS_MODELS_DIR, f"{NAME}-{timestamp}")
             sfs = SaturnFS()
             sfs.put(output_dir, remote_dir, recursive=True)
